Renames modules
* Renames dataframe_fetcher to df_fetcher
* Renames dataframe_repository to df_repository
